+++ src/main/java/com/raoulvdberge/refinedstorage/api/solderer/ISoldererRegistry.java
@@ -1,7 +1,5 @@
 package com.raoulvdberge.refinedstorage.api.solderer;
 
-import net.minecraft.item.ItemStack;
-import net.minecraft.util.NonNullList;
 import net.minecraftforge.items.IItemHandler;
 
 import javax.annotation.Nonnull;
@@ -32,15 +30,4 @@ public interface ISoldererRegistry {
      * @return a list with all the solderer recipes
      */
     List<ISoldererRecipe> getRecipes();
-
-    /**
-     * Creates a simple solderer recipe.
-     *
-     * @param result   the result
-     * @param duration the duration in ticks
-     * @param rows     the rows of this recipe, has to be 3 rows (empty list for empty row)
-     * @return a solderer recipe
-     */
-    @Nonnull
-    ISoldererRecipe createSimpleRecipe(@Nonnull ItemStack result, int duration, NonNullList<ItemStack>... rows);
 }