+++ src/main/java/com/raoulvdberge/refinedstorage/item/ItemUpgrade.java
@@ -2,19 +2,13 @@ package com.raoulvdberge.refinedstorage.item;
 
 import com.raoulvdberge.refinedstorage.RS;
 import com.raoulvdberge.refinedstorage.RSItems;
-import com.raoulvdberge.refinedstorage.RSUtils;
 import net.minecraft.client.resources.I18n;
 import net.minecraft.client.util.ITooltipFlag;
 import net.minecraft.creativetab.CreativeTabs;
-import net.minecraft.enchantment.Enchantment;
-import net.minecraft.enchantment.EnchantmentData;
-import net.minecraft.init.Items;
-import net.minecraft.item.ItemEnchantedBook;
 import net.minecraft.item.ItemStack;
 import net.minecraft.nbt.NBTTagCompound;
 import net.minecraft.util.NonNullList;
 import net.minecraft.world.World;
-import net.minecraftforge.oredict.OreDictionary;
 
 import javax.annotation.Nullable;
 import java.util.List;
@@ -103,25 +97,6 @@ public class ItemUpgrade extends ItemBase {
                 return RS.INSTANCE.config.fortuneUpgradeUsagePerFortune * ItemUpgrade.getFortuneLevel(stack);
             default:
                 return 0;
-        }
-    }
-
-    public static NonNullList<ItemStack> getRequirement(ItemStack stack) {
-        switch (stack.getItemDamage()) {
-            case ItemUpgrade.TYPE_RANGE:
-                return OreDictionary.getOres("enderpearl");
-            case ItemUpgrade.TYPE_SPEED:
-                return NonNullList.withSize(1, new ItemStack(Items.SUGAR));
-            case ItemUpgrade.TYPE_CRAFTING:
-                return OreDictionary.getOres("workbench");
-            case ItemUpgrade.TYPE_INTERDIMENSIONAL:
-                return OreDictionary.getOres("netherStar");
-            case ItemUpgrade.TYPE_SILK_TOUCH:
-                return NonNullList.withSize(1, ItemEnchantedBook.getEnchantedItemStack(new EnchantmentData(Enchantment.getEnchantmentByLocation("silk_touch"), 1)));
-            case ItemUpgrade.TYPE_FORTUNE:
-                return NonNullList.withSize(1, ItemEnchantedBook.getEnchantedItemStack(new EnchantmentData(Enchantment.getEnchantmentByLocation("fortune"), getFortuneLevel(stack))));
-            default:
-                return RSUtils.emptyNonNullList();
         }
     }
 }