+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### 1.5.9
 - Fixed crash relating to MCMP (raoulvdberge)
+- Fixed not being able to extract anything when connecting an External Storage to Storage Drawers (raoulvdberge)
 
 ### 1.5.8
 - Updated Forge to 2400 (raoulvdberge)